Modularised.
This makes the tests much easier to understand, and more thorough. The
template funtionality doesn't quite work yet, and there are probably other
glitches. We need some integration tests to help identify these.

I'd also like to do a bit of renaming and restructuring of the modules.
The name pretendr is littered everywhere for example - it'd be better if
it was a bit more brand-agnostic.